+++ package.json
@@ -13,6 +13,7 @@
     "@material-ui/icons": "^4.11.2",
     "axios": "^0.26.1",
     "bootstrap": "^5.1.3",
+    "bootstrap-icons": "^1.8.1",
     "next": "12.1.0",
     "node-sass": "^7.0.1",
     "qs": "^6.10.3",
