+++ styles/globals.css
@@ -18,7 +18,7 @@ a {
 
 .radio-btn {
   border-radius: 20px;
-
+}
 iframe {
   height: 100%;
   width: 100%;
